bugfix for supergroups

+++ migrations/20151024134130_users.php
@@ -9,7 +9,7 @@ class Users extends AbstractMigration
         $user_account = $this->table('users')
             ->addColumn('username', 'string', array('limit' => 150))
             ->addColumn('registered_by', 'integer')
-            ->addColumn('chat_id', 'integer')
+            ->addColumn('chat_id', 'bigint')
             ->addColumn('created', 'datetime', array('default' => 'CURRENT_TIMESTAMP'))
             ->addIndex(array('username', 'chat_id'), array('unique' => true));
         $user_account->save();
